adjust for Go 1.19 changes

+++ rlimit.go
@@ -2,6 +2,9 @@
 // Use of this source code is governed by a BSD-style
 // license that can be found in the LICENSE file.
 
+//go:build freebsd
+// +build freebsd
+
 package sqlite // import "modernc.org/sqlite"
 
 import (
@@ -10,7 +13,7 @@ import (
 
 func setMaxOpenFiles(n int64) error {
 	var rLimit unix.Rlimit
-	rLimit.Max = uint64(n)
-	rLimit.Cur = uint64(n)
+	rLimit.Max = n
+	rLimit.Cur = n
 	return unix.Setrlimit(unix.RLIMIT_NOFILE, &rLimit)
 }
